+++ policyengine_us/tests/policy/baseline/gov/irs/credits/ctc/ctc_reduction.yaml
@@ -1,31 +1,24 @@
-- name: TCJA reduction, high income
+- name: Head of household with $220k income has a $1,000 CTC reduction under TCJA.
   period: 2020
   input:
-    people:
-      child:
-        age: 4
-      parent:
-        age: 21
-        is_tax_unit_dependent: false
-    tax_units:
-      tax_unit:
-        members: [child, parent]
-        adjusted_gross_income: 220_000
+    adjusted_gross_income: 220_000
+    ctc_phase_out_threshold: 200_000 # HOH.
   output:
-    ctc_reduction: 1_000
+    ctc_reduction: 1_000 # $20,000 * 5%.
 
-- name: ARPA reduction, high income
+- name: Head of household with $400k income has CTC reduction capped at maximum under TCJA.
+  period: 2020
+  input:
+    adjusted_gross_income: 400_000
+    ctc_phase_out_threshold: 200_000 # HOH.
+  output:
+    ctc_reduction: 10_000 # $200k * 5%. Gets capped at maximum.
+
+- name: Head of household with $220k income and a child under six has a $2,600 CTC reduction under ARPA.
   period: 2021
   input:
-    people:
-      child:
-        age: 4
-      parent:
-        age: 21
-        is_tax_unit_dependent: false
-    tax_units:
-      tax_unit:
-        members: [child, parent]
-        adjusted_gross_income: 220_000
+    adjusted_gross_income: 220_000
+    ctc_phase_out_threshold: 200_000 # HOH.
+    ctc_arpa_reduction: 1_600
   output:
     ctc_reduction: 2_600
